+++ packages/fiori-app-sub-generator/src/fiori-app-generator/prompting.ts
@@ -304,9 +304,6 @@ export async function createUI5ApplicationPromptOptions(
         [ui5AppInquirerPromptNames.enableTypeScript]: {
             default: defaultPromptValues[ui5AppInquirerPromptNames.enableTypeScript]
         },
-        [ui5AppInquirerPromptNames.enableCodeAssist]: {
-            hide: true
-        },
         [ui5AppInquirerPromptNames.enableVirtualEndpoints]: {
             hide: service.capService?.capType === 'Java'
         }
@@ -408,6 +405,7 @@ function createOdataServicePromptOptions(options: OdataServiceInquirerOptions): 
         [odataServiceInquirerPromptNames.capProject]: {
             capSearchPaths: options.workspaceFolders ?? [],
             defaultChoice: options.capService?.projectPath,
+            useAutoComplete: getHostEnvironment() === hostEnvironment.cli,
             ...options.promptOptions?.capProject
         },
         [odataServiceInquirerPromptNames.capService]: {