+++ examples/ui-prompting-examples/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @sap-ux-private/ui-prompting-examples
 
+## 0.4.14
+
+### Patch Changes
+
+- Updated dependencies [26dde65]
+    - @sap-ux/ui-components@1.27.2
+    - @sap-ux/ui-prompting@0.5.13
+
+## 0.4.13
+
+### Patch Changes
+
+- Updated dependencies [e111d0d]
+    - @sap-ux/ui-components@1.27.1
+    - @sap-ux/ui-prompting@0.5.12
+
+## 0.4.12
+
+### Patch Changes
+
+- Updated dependencies [f9f119d]
+    - @sap-ux/ui-components@1.27.0
+    - @sap-ux/ui-prompting@0.5.11
+
+## 0.4.11
+
+### Patch Changes
+
+- Updated dependencies [87a6696]
+    - @sap-ux/ui-components@1.26.19
+    - @sap-ux/ui-prompting@0.5.10
+
 ## 0.4.10
 
 ### Patch Changes