+++ packages/odata-service-writer/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @sap-ux/odata-service-writer
 
+## 0.29.7
+
+### Patch Changes
+
+- Updated dependencies [55ac9f0]
+    - @sap-ux/project-access@1.34.1
+    - @sap-ux/mockserver-config-writer@0.9.39
+
+## 0.29.6
+
+### Patch Changes
+
+- Updated dependencies [b132944]
+    - @sap-ux/project-access@1.34.0
+    - @sap-ux/mockserver-config-writer@0.9.38
+
+## 0.29.5
+
+### Patch Changes
+
+- Updated dependencies [f13d1c2]
+    - @sap-ux/project-access@1.33.2
+    - @sap-ux/mockserver-config-writer@0.9.37
+
+## 0.29.4
+
+### Patch Changes
+
+- Updated dependencies [e111d0d]
+    - @sap-ux/project-access@1.33.1
+    - @sap-ux/mockserver-config-writer@0.9.36
+
+## 0.29.3
+
+### Patch Changes
+
+- 2204ad3: fix(deps): update dependencies @sap-ux/annotation-converter to v0.10.19 and @sap-ux/vocabularies-types to v0.14.5
+
+## 0.29.2
+
+### Patch Changes
+
+- Updated dependencies [c7f9a60]
+    - @sap-ux/project-access@1.33.0
+    - @sap-ux/mockserver-config-writer@0.9.35
+    - @sap-ux/ui5-config@0.29.12
+
 ## 0.29.1
 
 ### Patch Changes