+++ packages/reload-middleware/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/reload-middleware
 
+## 0.3.10
+
+### Patch Changes
+
+- ce4b29c: Upgrade qs/body-parser/express
+
+## 0.3.9
+
+### Patch Changes
+
+- Updated dependencies [c7f9a60]
+    - @sap-ux/logger@0.8.0
+    - @sap-ux/btp-utils@1.1.6
+
 ## 0.3.8
 
 ### Patch Changes