+++ packages/store/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/store
 
+## 1.5.0
+
+### Minor Changes
+
+- c9fd939: update backend systems with system info from adt api
+
+## 1.4.2
+
+### Patch Changes
+
+- Updated dependencies [c7f9a60]
+    - @sap-ux/logger@0.8.0
+
 ## 1.4.1
 
 ### Patch Changes