+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,89 @@
 # @sap-ux/ui5-application-inquirer
 
+## 0.16.7
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.10.12
+
+## 0.16.6
+
+### Patch Changes
+
+- Updated dependencies [55ac9f0]
+    - @sap-ux/project-access@1.34.1
+    - @sap-ux/project-input-validator@0.6.44
+    - @sap-ux/inquirer-common@0.10.11
+
+## 0.16.5
+
+### Patch Changes
+
+- Updated dependencies [b132944]
+    - @sap-ux/project-access@1.34.0
+    - @sap-ux/project-input-validator@0.6.43
+    - @sap-ux/inquirer-common@0.10.10
+
+## 0.16.4
+
+### Patch Changes
+
+- Updated dependencies [f13d1c2]
+    - @sap-ux/project-access@1.33.2
+    - @sap-ux/project-input-validator@0.6.42
+    - @sap-ux/inquirer-common@0.10.9
+
+## 0.16.3
+
+### Patch Changes
+
+- Updated dependencies [e111d0d]
+    - @sap-ux/inquirer-common@0.10.8
+    - @sap-ux/project-access@1.33.1
+    - @sap-ux/project-input-validator@0.6.41
+
+## 0.16.2
+
+### Patch Changes
+
+- Updated dependencies [2204ad3]
+    - @sap-ux/inquirer-common@0.10.7
+
+## 0.16.1
+
+### Patch Changes
+
+- Updated dependencies [6382440]
+    - @sap-ux/inquirer-common@0.10.6
+
+## 0.16.0
+
+### Minor Changes
+
+- 03598eb: Remove codeAssist support during project generation prompts and writing.
+
+### Patch Changes
+
+- Updated dependencies [03598eb]
+    - @sap-ux/ui5-info@0.13.7
+    - @sap-ux/inquirer-common@0.10.5
+
+## 0.15.53
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.10.4
+
+## 0.15.52
+
+### Patch Changes
+
+- Updated dependencies [c7f9a60]
+    - @sap-ux/project-access@1.33.0
+    - @sap-ux/inquirer-common@0.10.3
+    - @sap-ux/ui5-info@0.13.6
+    - @sap-ux/project-input-validator@0.6.40
+
 ## 0.15.51
 
 ### Patch Changes