+++ packages/ui5-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/ui5-proxy-middleware
 
+## 1.6.11
+
+### Patch Changes
+
+- ce4b29c: Upgrade qs/body-parser/express
+
+## 1.6.10
+
+### Patch Changes
+
+- Updated dependencies [c7f9a60]
+    - @sap-ux/logger@0.8.0
+    - @sap-ux/ui5-config@0.29.12
+
 ## 1.6.9
 
 ### Patch Changes